@@ -1,5 +1,5 @@
 {
-  "name": "angular-ng",
+  "name": "js-framework-benchmark-angular-ng",
   "version": "0.0.0",
   "js-framework-benchmark": {
     "frameworkVersionFromPackage": "@angular/core",
@@ -16,24 +16,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "9.1.2",
-    "@angular/common": "9.1.2",
-    "@angular/compiler": "9.1.2",
-    "@angular/core": "9.1.2",
-    "@angular/platform-browser": "9.1.2",
-    "@angular/platform-browser-dynamic": "9.1.2",
+    "@angular/animations": "10.0.4",
+    "@angular/common": "10.0.4",
+    "@angular/compiler": "10.0.4",
+    "@angular/core": "10.0.4",
+    "@angular/platform-browser": "10.0.4",
+    "@angular/platform-browser-dynamic": "10.0.4",
     "rxjs": "6.5.5",
-    "tslib": "^1.11.1",
+    "tslib": "^2.0.0",
     "zone.js": "0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.901.1",
-    "@angular/cli": "9.1.1",
-    "@angular/compiler-cli": "9.1.2",
-    "@angular/language-service": "9.1.2",
+    "@angular-devkit/build-angular": "0.1000.3",
+    "@angular/cli": "10.0.3",
+    "@angular/compiler-cli": "10.0.4",
+    "@angular/language-service": "10.0.4",
     "@types/node": "13.13.0",
     "ts-node": "8.8.2",
-    "tslint": "6.1.1",
-    "typescript": "3.8.3"
+    "tslint": "~6.1.0",
+    "typescript": "3.9.7"
   }
 }
