@@ -10,15 +10,14 @@
     "build-prod": "webpack -p"
   },
   "devDependencies": {
-    "babel-core": "6.26.3",
-    "babel-loader": "7.1.5",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-react": "6.24.1",
+    "@babel/core": "^7.11.6",
+    "babel-preset-atomic": "^2.4.2",
+    "babel-loader": "^8.1.0",
     "jsx-loader": "0.13.2",
-    "webpack": "4.16.3",
-    "webpack-cli": "^3.1.0"
+    "webpack": "4.44.2",
+    "webpack-cli": "^3.3.12"
   },
   "dependencies": {
-    "etch": "0.14.0"
+    "etch": "0.14.1"
   }
 }
