@@ -1,5 +1,5 @@
 {
-  "name": "js-framework-benchmark-react-lite",
+  "name": "js-framework-benchmark-preact",
   "version": "1.0.0",
   "description": "Benchmark for react-lite framework",
   "js-framework-benchmark": {
@@ -20,17 +20,17 @@
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
   "devDependencies": {
-    "babel-minify-webpack-plugin": "^0.3.1",
-    "@babel/core": "7.6.4",
-    "@babel/preset-env": "7.6.3",
-    "@babel/preset-react": "7.6.3",
-    "@babel/plugin-proposal-class-properties": "7.5.5",
-    "babel-loader": "8.0.6",
-    "webpack": "4.41.2",
-    "webpack-cli": "3.3.9"
+    "babel-minify-webpack-plugin": "0.3.1",
+    "@babel/core": "7.11.6",
+    "@babel/preset-env": "7.11.5",
+    "@babel/preset-react": "7.10.4",
+    "@babel/plugin-proposal-class-properties": "7.10.4",
+    "babel-loader": "8.1.0",
+    "webpack": "4.44.2",
+    "webpack-cli": "3.3.12"
   },
   "dependencies": {
-    "@babel/plugin-transform-react-jsx": "7.3.0",
-    "preact": "10.0.1"
+    "@babel/plugin-transform-react-jsx": "7.10.4",
+    "preact": "10.5.5"
   }
 }
