@@ -1,5 +1,5 @@
 {
-  "name": "js-framework-benchmark-react",
+  "name": "js-framework-benchmark-react-mobx",
   "version": "1.1.1",
   "description": "React demo",
   "main": "index.js",
@@ -30,13 +30,14 @@
     "babel-preset-react": "6.24.1",
     "babel-preset-stage-0": "6.24.1",
     "jsx-loader": "0.13.2",
+    "terser-webpack-plugin": "^3.0.6",
     "webpack": "4.16.3",
     "webpack-cli": "^3.1.0"
   },
   "dependencies": {
-    "mobx": "5.0.3",
-    "mobx-react": "5.2.3",
-    "react": "16.4.1",
-    "react-dom": "16.4.1"
+    "mobx": "5.15.4",
+    "mobx-react": "6.2.2",
+    "react": "17.0.1",
+    "react-dom": "17.0.1"
   }
 }
