@@ -1,5 +1,5 @@
 {
-  "name": "js-framework-benchmark-react",
+  "name": "js-framework-benchmark-react-redux",
   "version": "1.1.1",
   "description": "React demo",
   "main": "index.js",
@@ -22,19 +22,19 @@
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
   "devDependencies": {
-    "@babel/core": "7.4.5",
-    "@babel/preset-env": "7.4.5",
-    "@babel/preset-react": "7.0.0",
-    "@babel/plugin-proposal-class-properties": "7.4.4",
-    "babel-loader": "8.0.6",
+    "@babel/core": "7.12.3",
+    "@babel/plugin-proposal-class-properties": "7.12.1",
+    "@babel/preset-env": "7.12.1",
+    "@babel/preset-react": "7.12.1",
+    "babel-loader": "8.1.0",
     "terser-webpack-plugin": "1.3.0",
     "webpack": "4.34.0",
     "webpack-cli": "3.3.4"
   },
   "dependencies": {
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "react-redux": "7.1.0",
-    "redux": "4.0.1"
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
+    "react-redux": "7.2.1",
+    "redux": "4.0.5"
   }
 }
