@@ -10,16 +10,16 @@
   "author": "Utkarsh Kukreti",
   "license": "MIT",
   "dependencies": {
-    "react": "16.13.0",
-    "react-dom": "16.13.0",
-    "reaml": "0.13.0"
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
+    "reaml": "0.15.0"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "11.0.2",
-    "@rollup/plugin-node-resolve": "7.1.1",
-    "@rollup/plugin-replace": "2.3.1",
-    "bs-platform": "7.2.0",
-    "rollup": "1.32.0",
-    "rollup-plugin-terser": "5.2.0"
+    "@rollup/plugin-commonjs": "15.1.0",
+    "@rollup/plugin-node-resolve": "9.0.0",
+    "@rollup/plugin-replace": "2.3.3",
+    "bs-platform": "8.3.1",
+    "rollup": "2.32.1",
+    "rollup-plugin-terser": "7.0.2"
   }
 }
