@@ -1,4 +1,4 @@
-import { create, w, v } from '@dojo/framework/core/vdom'
+import { create, w, v, diffProperty } from '@dojo/framework/core/vdom'
 import Button from './Button';
 
 export interface ButtonConfig {
@@ -11,14 +11,15 @@ export interface ButtonsProperties {
 	buttonConfigs: ButtonConfig[];
 }
 
-const factory = create().properties<ButtonsProperties>();
+const factory = create({ diffProperty }).properties<ButtonsProperties>();
 
-export default factory(function Buttons({ properties }) {
+export default factory(function Buttons({ properties, middleware: { diffProperty } }) {
+	diffProperty('buttonConfigs', properties, () => {});
 	const { buttonConfigs } = properties();
 	return v('div', { classes: [ 'jumbotron' ] }, [
 		v('div', { classes: [ 'row' ] }, [
 			v('div', { classes: [ 'col-md-6' ] }, [
-				v('h1', ['Dojo v6.0.0'])
+				v('h1', ['Dojo v7'])
 			]),
 			v('div', { classes: [ 'col-md-6' ] }, buttonConfigs.map(({ id, label, onClick }) => {
 				return w(Button, { key: id, id, label, onClick });
