@@ -1,5 +1,5 @@
 'use strict';
 
-let {Main} = require('./Main');
-let component = new Main({});
+const {Main} = require('./Main');
+const component = new Main({});
 document.getElementById('main').appendChild(component.element);
