@@ -14,7 +14,5 @@ const minify = ( file ) => {
         return code
     }
 }
-
-fs.writeFileSync( 'dist/fntags.js', minify( 'node_modules/fntags/src/fntags.js' ) )
-
-fs.writeFileSync( 'dist/Main.js', minify( 'src/Main.js' ).replace( '../node_modules/fntags/src/fntags.js', './fntags.js' ) )
+fs.writeFileSync('dist/fntags.min.js', minify('node_modules/fntags/src/fntags.js'))
+fs.writeFileSync( 'dist/Main.js', minify( 'src/Main.js' ) )
