@@ -1,119 +1,159 @@
-import { h, Lazy, app } from "hyperapp"
-import { state, actions } from "./store"
+import { h, memo, text, app } from "hyperapp"
+import { buildData } from "./store"
 
-const Row = ({ data, label, styleClass }) => (
-  <tr key={data.id} class={styleClass}>
-    <td class="col-md-1">{data.id}</td>
-    <td class="col-md-4">
-        <a onclick={[actions.select, data.id]}>{label}</a>
-    </td>
-    <td class="col-md-1">
-        <a onclick={[actions.delete, data.id]}>
-            <span class="glyphicon glyphicon-remove" aria-hidden="true">
-            </span>
-        </a>
-    </td>
-    <td class="col-md-6"></td>
-  </tr>
-);
+const updateRow = (id, label) => ({
+  id,
+  label: label + " !!!",
+})
 
-const LazyRow = ({ data, label, styleClass }) => (
-  <Lazy
-    view={Row}
-    data={data}
-    label={label}
-    styleClass={styleClass}
-  />
-);
+const append1K = (state) => ({
+  data: state.data.concat(buildData(1000)),
+})
+
+const create1K = () => ({
+  data: buildData(1000),
+})
+
+const create10K = () => ({
+  data: buildData(10000),
+})
+
+const clearAllRows = () => ({
+  data: [],
+})
+
+const deleteRow = (state, id) => ({
+  data: state.data.filter((d) => d.id !== id),
+})
+
+const updateEveryTenth = (state) => ({
+  data: state.data.map((d, i) => (i % 10 === 0 ? updateRow(d.id, d.label) : d)),
+  selected: undefined,
+})
+
+const selectRow = (state, id) => ({
+  data: state.data,
+  selected: id,
+})
+
+const swapRows = (state) => {
+  if (state.data.length <= 998) return state
+
+  const temp = state.data[1]
+  state.data[1] = state.data[998]
+  state.data[998] = temp
+
+  return {
+    data: state.data,
+    selected: state.selected,
+  }
+}
+
+const Row = ({ data, label, styleClass }) =>
+  h("tr", { key: data.id, class: styleClass }, [
+    h("td", { class: "col-md-1" }, text(data.id)),
+    h("td", { class: "col-md-4" },
+      h("a", { onclick: [ selectRow, data.id ] }, text(label)),
+    ),
+    h("td", { class: "col-md-1" },
+      h("a", { onclick: [ deleteRow, data.id ] },
+        h("span", {
+          class: "glyphicon glyphicon-remove",
+          "aria-hidden": "true",
+        })
+      ),
+    ),
+    h("td", { class: "col-md-6" }),
+  ])
 
 app({
-  init: state,
-  view: state => (
-    <div class="container">
-      <div class="jumbotron">
-        <div class="row">
-          <div class="col-md-6">
-            <h1>HyperApp</h1>
-          </div>
-          <div class="col-md-6">
-            <div class="row">
-              <div class="col-sm-6 smallpad">
-                <button
-                  type="button"
-                  class="btn btn-primary btn-block"
-                  id="run"
-                  onclick={actions.run}
-                >
-                  Create 1,000 rows
-                </button>
-              </div>
-              <div class="col-sm-6 smallpad">
-                <button
-                  type="button"
-                  class="btn btn-primary btn-block"
-                  id="runlots"
-                  onclick={actions.runLots}
-                >
-                  Create 10,000 rows
-                </button>
-              </div>
-              <div class="col-sm-6 smallpad">
-                <button
-                  type="button"
-                  class="btn btn-primary btn-block"
-                  id="add"
-                  onclick={actions.add}
-                >
-                  Append 1,000 rows
-                </button>
-              </div>
-              <div class="col-sm-6 smallpad">
-                <button
-                  type="button"
-                  class="btn btn-primary btn-block"
-                  id="update"
-                  onclick={actions.update}
-                >
-                  Update every 10th row
-                </button>
-              </div>
-              <div class="col-sm-6 smallpad">
-                <button
-                  type="button"
-                  class="btn btn-primary btn-block"
-                  id="clear"
-                  onclick={actions.clear}
-                >
-                  Clear
-                </button>
-              </div>
-              <div class="col-sm-6 smallpad">
-                <button
-                  type="button"
-                  class="btn btn-primary btn-block"
-                  id="swaprows"
-                  onclick={actions.swapRows}
-                >
-                  Swap Rows
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <table class="table table-hover table-striped test-data">
-        <tbody>
-          {state.data.map(data => (
-            <LazyRow
-              data={data}
-              label={data.label}
-              styleClass={data.id === state.selected ? 'danger' : ''}
-            />
-          ))}
-        </tbody>
-      </table>
-      <span class="preloadicon glyphicon glyphicon-remove" aria-hidden="true" />
-    </div>
-  ),
-  node: document.getElementById('main')
-});
+  init: {
+    data: [],
+    selected: false,
+  },
+  view: (state) =>
+    h("div", { class: "container" }, [
+      h("div", { class: "jumbotron" },
+        h("div", { class: "row" }, [
+          h("div", { class: "col-md-6" }, h("h1", {}, text("Hyperapp"))),
+          h("div", { class: "col-md-6" },
+            h("div", { class: "row" }, [
+              h("div", { class: "col-sm-6 smallpad" },
+                h("button", {
+                    type: "button",
+                    class: "btn btn-primary btn-block",
+                    id: "run",
+                    onclick: create1K,
+                  },
+                  text("Create 1,000 rows")
+                )
+              ),
+              h("div", { class: "col-sm-6 smallpad" },
+                h("button", {
+                    type: "button",
+                    class: "btn btn-primary btn-block",
+                    id: "runlots",
+                    onclick: create10K,
+                  },
+                  text("Create 10,000 rows")
+                )
+              ),
+              h("div", { class: "col-sm-6 smallpad" },
+                h("button", {
+                    type: "button",
+                    class: "btn btn-primary btn-block",
+                    id: "add",
+                    onclick: append1K,
+                  },
+                  text("Append 1,000 rows")
+                )
+              ),
+              h("div", { class: "col-sm-6 smallpad" },
+                h("button", {
+                    type: "button",
+                    class: "btn btn-primary btn-block",
+                    id: "update",
+                    onclick: updateEveryTenth,
+                  },
+                  text("Update every 10th row")
+                )
+              ),
+              h("div", { class: "col-sm-6 smallpad" },
+                h("button", {
+                    type: "button",
+                    class: "btn btn-primary btn-block",
+                    id: "clear",
+                    onclick: clearAllRows,
+                  },
+                  text("Clear")
+                )
+              ),
+              h("div", { class: "col-sm-6 smallpad" },
+                h("button", {
+                    type: "button",
+                    class: "btn btn-primary btn-block",
+                    id: "swaprows",
+                    onclick: swapRows,
+                  },
+                  text("Swap Rows")
+                )
+              )
+            ])
+          )]
+        )
+      ),
+      h("table", { class: "table table-hover table-striped test-data" },
+        h("tbody", {}, state.data.map((data) => memo(Row, {
+              data,
+              label: data.label,
+              styleClass: data.id === state.selected ? "danger" : "",
+          }))
+        )
+      ),
+      h("span", {
+        class: "preloadicon glyphicon glyphicon-remove",
+        "aria-hidden": "true",
+      })
+    ]),
+  node: document.getElementById("app"),
+})
