@@ -1,14 +1,13 @@
 import Mikado from "../node_modules/mikado/src/mikado.js";
 import Array from "../node_modules/mikado/src/array.js";
-import app from "./template/app.es6.js";
-import item from "./template/item.es6.js";
-import { buildData } from "./data.js";
+import tpl_app from "./template/app.es6.js";
+import tpl_item from "./template/item.es6.js";
+import buildData from "./data.js";
 
-Mikado.once(document.getElementById("main"), app);
+Mikado.once(document.getElementById("main"), tpl_app);
 
-let selected = 0;
 const store = new Array();
-const view = new Mikado(document.getElementById("tbody"), item, {
+const view = new Mikado(document.getElementById("tbody"), tpl_item, {
     "reuse": false, "store": store
 })
 .route("run", () => store.set(buildData(1000)))
@@ -25,8 +24,4 @@ const view = new Mikado(document.getElementById("tbody"), item, {
     store[1] = tmp;
 })
 .route("remove", target => store.splice(view.index(target), 1))
-.route("select", target => {
-    store[selected]["selected"] = "";
-    store[selected = view.index(target)]["selected"] = "danger";
-})
-.listen("click");
+.route("select", target => view.refresh(view.selected = view.index(target)));
