@@ -52,9 +52,9 @@ const parameter = (function(opt){
     compilation_level: "ADVANCED_OPTIMIZATIONS", //"WHITESPACE"
     use_types_for_optimization: true,
     //new_type_inf: true,
-    jscomp_warning: "newCheckTypes",
+    //jscomp_warning: "newCheckTypes",
     //jscomp_error: "strictCheckTypes",
-    jscomp_error: "newCheckTypesExtraChecks",
+    //jscomp_error: "newCheckTypesExtraChecks",
     generate_exports: true,
     export_local_property_definitions: true,
     language_in: "ECMASCRIPT6_STRICT",
