@@ -1,5 +1,4 @@
 'use strict';
-/** @jsx preact.h */
 
 var preact = require('preact');
 var { render, h, Component } = preact;
@@ -10,6 +9,7 @@ export class Row extends Component {
 		this.onDelete = this.onDelete.bind(this);
 		this.onClick = this.onClick.bind(this);
 	}
+
 	shouldComponentUpdate(nextProps, nextState) {
 		return nextProps.data !== this.props.data || nextProps.styleClass !== this.props.styleClass;
 	}
@@ -17,12 +17,13 @@ export class Row extends Component {
 	onDelete() {
 		this.props.onDelete(this.props.data.id);
 	}
+	
 	onClick() {
 		this.props.onClick(this.props.data.id);
 	}
 
 	render() {
-		let {styleClass, onClick, onDelete, data} = this.props;
+		let {styleClass, data} = this.props;
 		return (<tr className={styleClass}>
 			<td className="col-md-1">{data.id}</td>
 			<td className="col-md-4">
