@@ -2,7 +2,6 @@ import { create, v } from '@dojo/framework/core/vdom';
 import store from './Store';
 
 export interface RowProperties {
-	onSelect: Function;
 	id: number;
 	label: string;
 }
@@ -10,15 +9,15 @@ export interface RowProperties {
 const factory = create({ store }).properties<RowProperties>();
 
 export default factory(function Row({ properties, middleware: { store } }) {
-	const { id, label, onSelect } = properties();
+	const { id, key = id, label } = properties();
 
 	return v('tr', {
-		classes: [ store.selected === id && 'danger' ]
+		classes: [ store.selected === key && 'danger' ]
 	}, [
 		v('td', { classes: [ 'col-md-1' ] }, [ `${id}` ]),
 		v('td', { classes: [ 'col-md-4' ] }, [
 			v('a', { onclick: () => {
-				onSelect(id);
+				store.select();
 			} }, [ label ])
 		]),
 		v('td', { classes: [ 'col-md-1' ] }, [
